The project is finally complete. closes #1
Included an executable file and made changes to the util.py file.
Earlier, due to testing purposes, the database did not commit any
changes made to it. This has been changed so that any changes made
are commited to the database. This should hopefully be the last
commit of this repo.